backend/requirements.txt
cryptography==45.0.3
Django==5.2
django-cors-headers==3.14.0
django-filter==23.4
django-jazzmin==3.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
drf-yasg==1.21.10
pillow==11.2.1
PyJWT==2.9.0
python-dotenv==1.1.0
requests==2.32.3
rest-framework-simplejwt==0.0.2
psycopg2-binary==2.9.9